# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Cuadra API
  version: "1.0.0"
  description: |
    REST API del backend de Cuadra POS.
    Multi-tenant: todos los endpoints filtran por `ownerId` del JWT.
    Auth: Bearer JWT en header `Authorization`.
servers:
  - url: http://localhost:4000
    description: Desarrollo local
  - url: https://cuadra-backend.railway.app
    description: Producción (Railway)

# ─── SECURITY ─────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        name: { type: string }
    TokenResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: bearer }
        expires_in: { type: integer, example: 604800 }
    UserProfile:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [owner, admin, cashier, viewer] }
        owner_id: { type: string, nullable: true }
        location_id: { type: string, nullable: true }
        commission_pct: { type: number, nullable: true }
        avatar_url: { type: string, nullable: true }

    # ── Products ────────────────────────────────────────────────────────────
...
    ClientCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        address: { type: string, nullable: true }
        notes: { type: string, nullable: true }
    ClientResponse:
      allOf:
        - $ref: "#/components/schemas/ClientCreate"
        - type: object
          properties:
            id: { type: string }
            owner_id: { type: string }
            total_debt: { type: number }
            active: { type: boolean }
            created_at: { type: string, format: date-time }

    # ── Expenses ─────────────────────────────────────────────────────────────
...
    CashboxCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        location_id: { type: string, nullable: true }
        assigned_user_id: { type: string, nullable: true }
    CashboxResponse:
      allOf:
        - $ref: "#/components/schemas/CashboxCreate"
        - type: object
          properties:
            id: { type: string }
            owner_id: { type: string }
            balance: { type: number }
            active: { type: boolean }

    # ── Cash Sessions ────────────────────────────────────────────────────────
    OpenSessionRequest:
      type: object
      properties:
        cashbox_id: { type: string, nullable: true }
        cashbox_name: { type: string, nullable: true }
    CloseSessionRequest:
      type: object
      properties:
        notes: { type: string, nullable: true }
        discrepancies: { type: string, nullable: true }
    CashSessionResponse:
      type: object
      properties:
        id: { type: string }
        owner_id: { type: string }
        cashier_id: { type: string }
        cashier_name: { type: string }
        cashbox_id: { type: string, nullable: true }
        status: { type: string, enum: [open, closed] }
        opened_at: { type: string, format: date-time }
        closed_at: { type: string, format: date-time, nullable: true }
        total_sales: { type: number }
        total_by_method:
          type: object
          properties:
            cash: { type: number }
            transfer: { type: number }
            mobile_pay: { type: number }
            credit: { type: number }
        debt_collected: { type: number }
        debt_pending_at_open: { type: number }
        debt_pending_at_close: { type: number }

    # ── Locations ────────────────────────────────────────────────────────────
...
security:
  - bearerAuth: []

# ─── PATHS ────────────────────────────────────────────────────────────────────
paths:
...
  /auth/register:
    post:
      summary: Registrar nuevo owner
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "400":
          description: Email ya registrado
  /auth/login:
    post:
      summary: Login — devuelve JWT
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Token JWT
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401":
          description: Credenciales inválidas
...
  /clients:
    get:
      summary: Listar clientes activos
      tags: [clients]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ClientResponse" }
          description: OK
    post:
      summary: Crear cliente
      tags: [clients]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientResponse" }
          description: Cliente creado
  /clients/{clientId}:
    parameters:
      - name: clientId
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Obtener cliente por ID
      tags: [clients]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientResponse" }
          description: OK
    patch:
      summary: Actualizar cliente
      tags: [clients]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientCreate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientResponse" }
          description: OK
    delete:
      summary: Soft-delete cliente (bloquea si tiene deudas)
      tags: [clients]
      responses:
        "204":
          description: Eliminado
        "400":
          description: Cliente con deudas pendientes

  # ── Expenses ─────────────────────────────────────────────────────────────────
...
  /cashboxes:
    get:
      summary: Listar cajas
      tags: [cashboxes]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CashboxResponse" }
          description: OK
    post:
      summary: Crear caja
      tags: [cashboxes]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CashboxCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CashboxResponse" }
          description: Caja creada
  /cashboxes/{cashboxId}:
    parameters:
      - name: cashboxId
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: Actualizar caja
      tags: [cashboxes]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CashboxCreate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CashboxResponse" }
          description: OK
    delete:
      summary: Desactivar caja
      tags: [cashboxes]
      responses:
        "204":
          description: Desactivada

  # ── Cash Sessions ────────────────────────────────────────────────────────────
  /cash-sessions/open:
    post:
      summary: Abrir sesión de caja
      tags: [cash-sessions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OpenSessionRequest" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CashSessionResponse" }
          description: Sesión abierta
        "400":
          description: Ya existe una sesión abierta para esta caja
  /cash-sessions/current:
    get:
      summary: Sesión activa del usuario autenticado
      tags: [cash-sessions]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CashSessionResponse" }
          description: Sesión activa
        "404":
          description: Sin sesión abierta
  /cash-sessions/{sessionId}/close:
    post:
      summary: Cerrar sesión de caja
      tags: [cash-sessions]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CloseSessionRequest" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CashSessionResponse" }
          description: Sesión cerrada
...